+++ packages/algob/src/builtin-tasks/gen-accounts.ts
@@ -1,9 +1,9 @@
+import { BuilderError, ERRORS } from "@algo-builder/web";
 import algosdk from "algosdk";
 import * as _fs from "fs";
 import * as path from "path";
 import YAML from "yaml";
 
-import { BuilderError, ERRORS } from "../errors/errors";
 import { task } from "../internal/core/config/config-env";
 import * as types from "../internal/core/params/argument-types";
 import { assertAllDirs, ASSETS_DIR } from "../internal/core/project-structure";
