+++ packages/algob/src/internal/cli/arguments-parser.ts
@@ -1,5 +1,5 @@
-import { BuilderError } from "../../errors/errors";
-import { ERRORS } from "../../errors/errors-list";
+import { BuilderError, ERRORS } from "@algo-builder/web";
+
 import {
   ParamDefinition,
   ParamDefinitions,
