+++ packages/algob/src/internal/cli/cli.ts
@@ -2,13 +2,12 @@
 // -*- mode: typescript -*- // https://github.com/syl20bnr/spacemacs/issues/13715
 import "source-map-support/register";
 
+import { BuilderError, BuilderPluginError, ERRORS } from "@algo-builder/web";
 import chalk from "chalk";
 import debug from "debug";
 import semver from "semver";
 
-import { TASK_HELP, TASK_INIT } from "../../builtin-tasks/task-names";
-import { BuilderError, BuilderPluginError } from "../../errors/errors";
-import { ERRORS } from "../../errors/errors-list";
+import { TASK_HELP } from "../../builtin-tasks/task-names";
 import { checkAlgorandUnauthorized } from "../../lib/exceptions";
 import { RuntimeArgs, RuntimeEnv, TaskArguments } from "../../types";
 import { ALGOB_NAME } from "../constants";
