+++ packages/algob/src/internal/core/plugins.ts
@@ -1,9 +1,8 @@
+import { BuilderError, ERRORS } from "@algo-builder/web";
 import debug from "debug";
 import * as path from "path";
 import * as semver from "semver";
 
-import { BuilderError } from "../../errors/errors";
-import { ERRORS } from "../../errors/errors-list";
 import { StrMap } from "../../types";
 import { BuilderContext } from "../context";
 import { ExecutionMode, getExecutionMode } from "./execution-mode";
