+++ packages/algob/test/builtin-tasks/deploy.ts
@@ -1,9 +1,9 @@
+import { ERRORS } from "@algo-builder/web";
 import { assert } from "chai";
 import * as fs from "fs";
 
 import { executeDeployTask as executeDeployTaskNoCLI, loadFilenames } from "../../src/builtin-tasks/deploy";
 import { TASK_DEPLOY, TASK_RUN } from "../../src/builtin-tasks/task-names";
-import { ERRORS } from "../../src/errors/errors-list";
 import { loadCheckpoint } from "../../src/lib/script-checkpoints";
 import { expectBuilderErrorAsync } from "../helpers/errors";
 import { testFixtureOutputFile, useCleanFixtureProject } from "../helpers/project";
