+++ packages/algob/test/fixture-projects/nested-node-project/top-caller-package-tester.ts
@@ -1,4 +1,5 @@
-import { getClosestCallerPackage } from "../../../src/internal/util/caller-package";
+import { getClosestCallerPackage } from "@algo-builder/web";
+
 import {
   call as callFromNested,
   callFromTopModule as nestedCallFromTopModule
