+++ packages/algob/test/internal/core/tasks/dsl.ts
@@ -1,6 +1,6 @@
+import { ERRORS } from "@algo-builder/web";
 import { assert } from "chai";
 
-import { ERRORS } from "../../../../src/errors/errors-list";
 import { TasksDSL } from "../../../../src/internal/core/tasks/dsl";
 import { expectBuilderErrorAsync } from "../../../helpers/errors";
