+++ packages/runtime/src/index.ts
@@ -1,6 +1,5 @@
 import { AccountStore } from "./account";
 import * as ERRORS from "./errors/errors-list";
-import { applyErrorMessageTemplate } from "./errors/runtime-errors";
 import { parseZodError } from "./errors/validation-errors";
 import { Interpreter } from "./interpreter/interpreter";
 import { loadASAFile, overrideASADef, validateASADefs } from "./lib/asa";
@@ -16,7 +15,6 @@ export {
   Runtime,
   AccountStore,
   checkIfAssetDeletionTx,
-  applyErrorMessageTemplate,
   loadFromYamlFileSilent,
   loadFromYamlFileSilentWithMessage,
   loadASAFile,
