+++ packages/web/src/index.ts
@@ -2,3 +2,9 @@ export * as types from "./types";
 export * as parsing from "./lib/parsing";
 export * as tx from "./lib/txn";
 export { ASADefSchema, ASADefsSchema } from "./types-input";
+export {
+  BuilderError, BuilderPluginError,
+  applyErrorMessageTemplate, parseAlgorandError
+} from "./errors/errors";
+export { ERRORS, ErrorDescriptor, ERROR_RANGES } from "./errors/errors-list";
+export { getClosestCallerPackage } from "./util/caller-package";
